requirements.txt
ftfy==6.1.3
horovod==0.28.1
huggingface_hub==0.13.4
matplotlib==3.5.3
matplotlib==3.4.3
numpy==1.21.6
opencv_python==4.6.0.66
opencv_python_headless==4.8.0.76
pandas==1.3.5
Pillow==9.2.0
Pillow==10.2.0
regex==2022.10.31
safetensors==0.2.8
scikit_image==0.19.3
scikit_learn==1.0.2
scipy==1.7.3
setuptools==60.2.0
setuptools==65.6.3
tabulate==0.9.0
timm==0.6.12
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.2
transformers==4.15.0
